Refactor sign in page object
- Move out into its parent directory as it's both support and publish
