+++ spec/support/page_objects/sign_in.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module PageObjects
+  class SignIn < PageObjects::Base
+    set_url "/sign-in"
+
+    element :page_heading, ".govuk-heading-l"
+
+    element :email_field, "#user-email-field"
+
+    element :sign_in_button, ".qa-sign_in_button"
+  end
+end
